adds the about us page and content, also adds the testing for it, and improved changes to the carousel components files based on feedback
