+++ app/javascript/components/components/Items.js
@@ -1,0 +1,64 @@
+  const items= [
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/1.svg',
+      altText: 'bulbasaur',
+      caption: 'bulbasaur',
+      key: 1,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/2.svg',
+      altText: 'ivysaur',
+      caption: 'ivysaur',
+      key: 2,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/3.svg',
+      altText: 'venasaur',
+      caption: 'venasaur',
+      key: 3,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/4.svg',
+      altText: 'charmander',
+      caption: 'charmander',
+      key: 4,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/5.svg',
+      altText: 'charmeleon',
+      caption: 'charmeleon',
+      key: 5,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/6.svg',
+      altText: 'charizard',
+      caption: 'charizard',
+      key: 6,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/7.svg',
+      altText: 'squirtle',
+      caption: 'squirtle',
+      key: 7,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/8.svg',
+      altText: 'wartortle',
+      caption: 'wartortle',
+      key: 8,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/9.svg',
+      altText: 'blastoise',
+      caption: 'blastoise',
+      key: 9,
+    },
+    {
+      src: 'https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/dream-world/25.svg',
+      altText: 'pikachu',
+      caption: 'pikachu',
+      key: 25,
+    },
+  ];
+
+export default items;
